+++ lib/web_translate_it/translation_file.rb
@@ -1,5 +1,4 @@
 module WebTranslateIt
-  
   # A TranslationFile is the representation of a master language file
   # on Web Translate It.
   #